requirements.txt
alembic==1.13.3
annotated-types==0.7.0
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
pluggy==1.5.0
prometheus_client==0.21.0
prometheus_flask_exporter==0.23.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
redis==5.2.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
